rename scannability_qreader
